ryzen-9-5950x-128gb-docker-zeus: Added build of individual componenets for Ryzen 9 5950x
This is created to be able to compare the Corsair MP600 (PCIe 4.0 NVMe)
to tmpfs. Notice the ext4 filesystem on the MP600 is not optimized.

The NVMe runs in a Ubuntu 18.04 docker with

docker container run -t -i --rm --privileged \
 --device /dev/net/tun:/dev/net/tun \
 -h p1_yocto2020.3 \
 -v /home/frosteyes/Desktop/oe-test/:/home/yocto/oe-test \
 --name yocto yocto-ubuntu-18.04-dev

The tmpfs version is build with the same docker image but uses 80GB for
tmp folder

docker container run -t -i --rm --privileged \
 --device /dev/net/tun:/dev/net/tun -h p1_yocto2020.3 \
 -v /home/frosteyes/Desktop/oe-test/:/home/yocto/oe-test \
 --tmpfs /home/yocto/oe-test/test-oe-build-time/poky/build/tmp:rw,suid,dev,exec,size=80G \
 --name yocto yocto-ubuntu-18.04-dev

Signed-off-by: Claus Stovgaard <claus.stovgaard@gmail.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>