results: add threadripper-7985wx-256gb-ubuntu-24.04-walnascar-2025-05
from @meta-gir

It's very impressive, the walnascar build times on my TR 3970x with 128GB ram
are around a double of this (and it's not just the double the number of cores
because 3990x wasn't twice as faster as my 3970x back in zeus build times).

Unfortunately we cannot easily compare with threadripper-3990x-zeus, because
a lot changed since zeus and e.g. newer chromium version takes much longer to
build compared to old one in zeus.